requirements.txt
Flask==2.2.3
openai==0.27.0
langchain==0.0.200
python-dotenv==1.0.0
